map remediation and remediator 1 to 1 & clean idx public types
