+++ lib/idx/index.ts
@@ -1,5 +1,4 @@
 export * from './authenticate';
-export * from './introspect';
 export * from './cancel';
 export * from './register';
 export * from './recoverPassword';
