+++ lib/idx/remediators/ChallengeAuthenticator.ts
@@ -1,0 +1,8 @@
+import { VerifyAuthenticator, VerifyAuthenticatorValues } from './Base/VerifyAuthenticator';
+
+export type ChallengeAuthenticatorValues = VerifyAuthenticatorValues;
+
+export class ChallengeAuthenticator extends VerifyAuthenticator {
+  static remediationName = 'challenge-authenticator';
+  values: ChallengeAuthenticatorValues;
+}
