+++ lib/idx/remediators/Identify.ts
@@ -1,4 +1,4 @@
-import { Remediator, RemediationValues } from './Remediator';
+import { Remediator, RemediationValues } from './Base/Remediator';
 
 export interface IdentifyValues extends RemediationValues {
   username?: string;
