+++ lib/idx/remediators/ResetAuthenticator.ts
@@ -1,0 +1,8 @@
+import { VerifyAuthenticator, VerifyAuthenticatorValues } from './Base/VerifyAuthenticator';
+
+export type ResetAuthenticatorValues = VerifyAuthenticatorValues;
+
+export class ResetAuthenticator extends VerifyAuthenticator {
+  static remediationName = 'reset-authenticator';
+  values: ResetAuthenticatorValues;
+}
