+++ lib/idx/remediators/SelectEnrollProfile.ts
@@ -1,9 +1,11 @@
-import { Remediator, RemediationValues } from './Remediator';
+import { Remediator, RemediationValues } from './Base/Remediator';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface SelectEnrollProfileValues extends RemediationValues {}
 
 export class SelectEnrollProfile extends Remediator {
+  static remediationName = 'select-enroll-profile';
+
   values: SelectEnrollProfileValues;
 
   canRemediate() {
