+++ lib/idx/remediators/index.ts
@@ -1,10 +1,13 @@
-export * from './Remediator';
-export * from './EnrollOrChallengeAuthenticator';
+export * from './Base/Remediator';
+export * from './EnrollAuthenticator';
+export * from './ChallengeAuthenticator';
+export * from './ResetAuthenticator';
 export * from './EnrollProfile';
 export * from './Identify';
 export * from './ReEnrollAuthenticator';
 export * from './RedirectIdp';
-export * from './SelectAuthenticator';
+export * from './SelectAuthenticatorAuthenticate';
+export * from './SelectAuthenticatorEnroll';
 export * from './SelectEnrollProfile';
 export * from './AuthenticatorVerificationData';
 export * from './AuthenticatorEnrollmentData';
