--- lib/idx/types.ts
@@ -1,160 +1,0 @@
-import { IdxTransactionMeta } from '../types/Transaction';
-import { Remediator } from './remediators';
-import { APIError, Tokens } from '../types';
-
-export { RemediationValues } from './remediators';
-export { AuthenticationOptions } from './authenticate';
-export { RegistrationOptions } from './register';
-export { PasswordRecoveryOptions } from './recoverPassword';
-export { CancelOptions } from './cancel';
-
-export type RemediationFlow = Record<string, typeof Remediator>;
-
-// A map from IDX data values (server spec) to RemediationValues (client spec)
-export type IdxToRemediationValueMap = Record<string, string[]>;
-
-export enum IdxStatus {
-  SUCCESS,
-  PENDING,
-  FAILURE,
-  TERMINAL,
-}
-
-type Input = {
-  name: string;
-  required?: boolean;
-}
-
-export type NextStep = {
-  name: string;
-  type?: string;
-  canSkip?: boolean;
-  inputs?: Input[];
-}
-
-export enum IdxFeature {
-  PASSWORD_RECOVERY,
-  REGISTRATION,
-  SOCIAL_IDP,
-}
-
-export interface IdxTransaction {
-  status: IdxStatus;
-  tokens?: Tokens;
-  nextStep?: NextStep;
-  messages?: IdxMessage[];
-  error?: APIError;
-  meta?: IdxTransactionMeta;
-  enabledFeatures?: IdxFeature[];
-  availableSteps?: NextStep[];
-}
-
-export interface AcceptsInteractionHandle {
-  interactionHandle?: string;
-}
-
-export interface IntrospectOptions extends AcceptsInteractionHandle {
-  stateHandle?: string;
-}
-
-export interface InteractOptions extends AcceptsInteractionHandle {
-  state?: string;
-  scopes?: string[];
-}
-
-export interface InteractResponse {
-  state?: string;
-  stateHandle?: string;
-  interactionHandle?: string;
-  idxResponse?: IdxResponse;
-  meta?: IdxTransactionMeta;
-}
-
-export type IdxOptions = InteractOptions;
-
-export interface IdpConfig {
-  id: string;
-  name: string;
-}
-
-// TODO: remove when idx-js provides type information
-export interface IdxAuthenticatorMethod {
-  type: string;
-}
-export interface IdxAuthenticator {
-  displayName: string;
-  id: string;
-  key: string;
-  methods: IdxAuthenticatorMethod[];
-  type: string;
-}
-
-export interface IdxForm {
-  value: IdxRemediationValue[];
-}
-
-export interface IdxOption {
-  value: string | { form: IdxForm };
-  label: string;
-  relatesTo?: IdxAuthenticator;
-}
-export interface IdxRemediationValue {
-  name: string;
-  type?: string;
-  required?: boolean;
-  secret?: boolean;
-  value?: string;
-  label?: string;
-  form?: IdxForm;
-  options?: IdxOption[];
-}
-export interface IdxRemediation {
-  name: string;
-  label?: string;
-  value: IdxRemediationValue[];
-  relatesTo?: {
-    type?: string;
-    value: IdxAuthenticator;
-  };
-  idp?: IdpConfig;
-  href?: string;
-  method?: string;
-  type?: string;
-}
-
-export interface IdxMessage {
-  message: string;
-  class: string;
-  i18n: {
-    key: string;
-  };
-}
-
-export interface IdxMessages {
-  type: string; // "array"
-  value: IdxMessage[];
-}
-
-// JSON response from the server
-export interface RawIdxResponse {
-  version: string;
-  stateHandle: string;
-  remediation?: IdxRemediation[];
-  messages?: IdxMessages;
-}
-
-export function isRawIdxResponse(obj: any): obj is RawIdxResponse {
-  return obj && obj.version;
-}
-
-// Object returned from idx-js
-export interface IdxResponse {
-  proceed: (remediationName: string, params: unknown) => Promise<IdxResponse>;
-  neededToProceed: IdxRemediation[];
-  rawIdxState: RawIdxResponse;
-  interactionCode?: string;
-  actions: Record<string, Function>;
-  toPersist: {
-    interactionHandle?: string;
-  };
-}
