+++ lib/idx/types/AcceptsInteractionHandle.ts
@@ -1,0 +1,3 @@
+export interface AcceptsInteractionHandle {
+  interactionHandle?: string;
+}
